HITrack/requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
django-filter>=23.5
drf-spectacular>=0.27.1
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests
pyyaml
psycopg2-binary==2.9.10
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
drf-nested-routers==0.94.1
django-filter
redis
gunicorn